+++ pdfs/dalitz/src/DPBWResonanceShape.cc
@@ -10,7 +10,7 @@ DPBWResonanceShape::DPBWResonanceShape(double mRR, double gammaRR, int L, double
 	,m2(mm2)
 {
 	pR0=DPHelpers::daughterMomentum(mR,m1,m2);
-	barrier = DPBarrierFactor(LR,RR,pR0);
+	barrier = DPBarrierFactor((unsigned int)LR,RR,pR0);
 }
 
 DPBWResonanceShape::DPBWResonanceShape( const DPBWResonanceShape& other ) : DPMassShape(other)
